+++ actiontext/lib/action_text/content.rb
@@ -97,8 +97,10 @@ module ActionText
 
     def render_attachments(**options, &block)
       content = fragment.replace(ActionText::Attachment.tag_name) do |node|
-        if node.key? "content"
+        if node.key?("content") && node["content"].present?
           node["content"] = sanitize_content_attachment(node["content"])
+        elsif node.key?("content")
+          node.remove_attribute("content")
         end
         block.call(attachment_for_node(node, **options))
       end