+++ actiontext/CHANGELOG.md
@@ -2,4 +2,8 @@
 
     *Petrik de Heus*
 
+*   Strip `content` attribute if the key is present but the value is empty
+
+    *Jeremy Green*
+
 Please check [7-2-stable](https://github.com/rails/rails/blob/7-2-stable/actiontext/CHANGELOG.md) for previous changes.